Code Smell: Switch Statement
The problem with the switch statement was that we could easily do this with a inheritance.

Each converter has a methodName() for showing the options, and it also has convert to give access to the converter to process the command. Thus there was no longer a need for a switch.